+++ ctpbee/__init__.py
@@ -16,3 +16,5 @@ from ctpbee.signals import send_monitor, cancel_monitor
 from ctpbee.trade_time import TradingDay
 from ctpbee.util import RiskLevel
 from ctpbee.func import hickey
+
+from ctpbee.looper.vessel import LooperApi, Vessel
